+++ core/client/components/error-boundary.js
@@ -1,6 +1,4 @@
-'use client';
-
-import React from "react";
+import React from 'react';
 
 export class ErrorBoundary extends React.Component {
   constructor(props) {
