+++ core/client/exports.js
@@ -1,0 +1,5 @@
+'use client';
+
+export { Link } from './components/link.js';
+export { useRouter } from './components/router/index.js';
+export { useFlash } from './components/flash-provider.js';
