+++ core/client/index.js
@@ -1,10 +1,10 @@
 import { hydrateRoot } from 'react-dom/client';
 import { createFromReadableStream } from 'react-server-dom-webpack/client';
-import { ErrorBoundary } from './error-boundary';
+import { ErrorBoundary } from './components/error-boundary';
 import { rscStream } from '../rsc-html-stream/client';
-import { SPARouter } from './spa-router';
-import { callServer } from './call-server';
-import { getFullPath } from './utils';
+import { SPARouter } from './components/router/spa-router';
+import { callServer } from './lib/call-server';
+import { getFullPath } from './lib/utils';
 
 async function hydrateDocument() {
   const { tree, formState, flashMessages } = await createFromReadableStream(
