+++ core/loaders/jsx.js
