+++ core/rsc-html-stream/client.js
