+++ core/server/exports.js
@@ -1,0 +1,4 @@
+const cookies = require('./lib/cookies');
+const flash = require('./lib/flash');
+
+module.exports = { cookies, flash };
