+++ core/server/lib/app-store.js
