+++ core/server/lib/flash.js
@@ -2,6 +2,5 @@ const { getAppStore } = require('./app-store');
 
 module.exports = function (...messages) {
   const { flashMessages } = getAppStore();
-
   flashMessages.push(...messages);
 };
