--- framework/client/exports.js
@@ -1,1 +1,0 @@
-export { useFlash } from './flash-context.js';
