--- framework/lib/exports.js
@@ -1,4 +1,0 @@
-const cookies = require('./cookies');
-const flash = require('./flash');
-
-module.exports = { cookies, flash };
